WIP: semi-working devcards
Uses the devcards iframe fork (though this is not actually necessary for
deftest stuff).

Two tests do not run without significant hacks. This is because these
tests use the [:create-page] event, which modified the URI hash and
causes all sorts of trouble. Need to rethink how this kind of navigation works...